\name{dtPosidonia-package}
\alias{dtPosidonia-package}
\alias{dtPosidonia}
\docType{package}
\title{
  Model of seagrass growth in aquatic environments, patterned to Posidonia
}
\description{
  Contains dynamic descriptions of the growth of Posidonia oceanica in the environment.
  
  The model is implemented in fortran and linked to R.
  
    \if{html}{\figure{Logo_posidonia.png}{options: width=80 alt="Logo" align='right'}}
\if{latex}{\figure{Logo_posidonia.jpg}{options: width=0.4in}}
  
}

\author{
  Karline Soetaert, Quiten Mudde, Joran de Gang, Marilaure Gregoire
}

\keyword{ package }
